Split tests across files, various style fixes
