+++ src/index.ts
@@ -2,8 +2,8 @@ import "angular";
 import "angular-cookies";
 import "angular-utf8-base64";
 
-import {NgJwtAuthServiceProvider} from "./ngJwtAuthServiceProvider";
-import {NgJwtAuthInterceptor} from "./ngJwtAuthInterceptor";
+import {NgJwtAuthServiceProvider} from "./provider/ngJwtAuthServiceProvider";
+import {NgJwtAuthInterceptor} from "./interceptor/ngJwtAuthInterceptor";
 
 angular.module('ngJwtAuth', ['utf8-base64', 'ngCookies'])
     .provider('ngJwtAuthService', NgJwtAuthServiceProvider)
