+++ src/test.ts
@@ -1,5 +1,5 @@
 // this file is only being used by karma
-require('phantomjs-polyfill')
+require('phantomjs-polyfill');
 
 requireAll((<any>require).context("./", true, /spec.ts$/));
 function requireAll(r: any): any {
